challenge_1b/requirements.txt
PyMuPDF>=1.22.0
sentence-transformers>=2.2.2
transformers>=4.30.0
faiss-cpu>=1.7.3
numpy>=1.21.0
tqdm>=4.64.0
nltk>=3.8.1